requirements.txt
Package                  Version
------------------------ ----------
aidial-sdk               0.17.0
aiohappyeyeballs         2.4.4
aiohttp                  3.11.11
aiosignal                1.3.2
annotated-types          0.7.0
anyio                    4.8.0
attrs                    24.3.0
certifi                  2024.12.14
charset-normalizer       3.4.1
click                    8.1.8
dataclasses-json         0.6.7
distro                   1.9.0
fastapi                  0.115.7
frozenlist               1.5.0
h11                      0.14.0
httpcore                 1.0.7
httpx                    0.28.1
httpx-sse                0.4.0
idna                     3.10
Jinja2                   3.1.5
jiter                    0.8.2
jsonpatch                1.33
jsonpointer              3.0.0
langchain                0.3.15
langchain-community      0.3.15
langchain-core           0.3.31
langchain-openai         0.3.2
langchain-text-splitters 0.3.5
langgraph                0.2.67
langgraph-checkpoint     2.0.10
langgraph-sdk            0.1.51
langsmith                0.3.1
MarkupSafe               3.0.2
marshmallow              3.26.0
msgpack                  1.1.0
multidict                6.1.0
mypy-extensions          1.0.0
numpy                    2.2.2
openai                   1.60.0
orjson                   3.10.15
packaging                24.2
pip                      24.3.1
propcache                0.2.1
pydantic                 2.10.6
pydantic_core            2.27.2
pydantic-settings        2.7.1
python-dotenv            1.0.1
PyYAML                   6.0.2
regex                    2024.11.6
requests                 2.32.3
requests-toolbelt        1.0.0
scrapingant-client       2.1.0
sniffio                  1.3.1
SQLAlchemy               2.0.37
starlette                0.45.3
tenacity                 9.0.0
tiktoken                 0.8.0
tqdm                     4.67.1
typing_extensions        4.12.2
typing-inspect           0.9.0
urllib3                  2.3.0
uvicorn                  0.34.0
wrapt                    1.17.2
yarl                     1.18.3
zstandard                0.23.0
